Verification tests (add Patlak to CPUfit), T1 exponential, CMake updates, CI stuff (#6)
* updated 7z command in create package

* Fixed CUBLAS linking on linux, static and dynamic

* updated version in linus python install script

* Added scripts for binary release, install, and testing  in linux

* added linux python install script

* updated file permissions on copy for linux python install

* use new cmake cuda tools, requires cmake 3.18

replaced depricated FindCuda with new LANGUAGES CUDA
simplifies architecture and cublas

* rename cmake latest model vars

* patlak added to cpu, trying to add gpu vs cpu tests

* Create c-cpp.yml for pipeline (#2)

Probably final. Need to investigate runner GPU otherwise we can only test compilation/build and CPU fitting.

* Update c-cpp.yml

* Expand Patlak cpu/gpu comparison

* Add boost tests to pipeline

* Update c-cpp.yml

stupid mistakes man

* undo boost tests

* add patlak performance comparison

* reformat Patlak comparison

* reformat Patlak comparison

* add boost to pipeline (#3)

* Create c-cpp.yml for pipeline (#2)

Probably final. Need to investigate runner GPU otherwise we can only test compilation/build and CPU fitting.

* fix patlak comparison stats, add cmake option for REAL to be double + improve data type compatibility

* force patlak comparison outputs to be double to mitigate precision error

* Patlak comparison formatting fixes + add convergence ratio

* add T1 fa exponential model, start on fitting example

* add t1 exponential option

* T1 Exponential WIP - runs but bad convergence

* add T1 FA exponential example

* Update c-cpp.yml

* automated build with github actions (Windows only)

* source folder is current directory

* don't install visual_studio_integration subpackage on linux

* call python not python.exe for cross-platform support

* linux build improvements

* fix linux Configure CMake

* typo

* Linux build command is simply make

* add conditionals

* create build directory

* fix paths

* linux cmake -DCMAKE_BUILD_TYPE=RELEASE

* upload entire build folder

* separate steps for windows and linux artifacts

* fixed Linux artifact upload

* created .deb file for linux install

* add cuda include dir to c++ compiles

* attempt to fix static linking with msvc

* added more version of cuda and ubuntu to test matrix

* remove quotes in matrix

* move cuda matrix

* indentation errors

* valid cuda-os combinations

* make pip verbose

* cmake output python location

* fix find python in cmake, drop old cuda from ubuntu

replace depricated find pythong with new function

* found another old python reference

* delete redundant message

* use jimkring/action-install-cuda-toolkit fork

* Added cuda 11.3.0

* use jimkring/action-install-cuda-toolkit fork

* test latest 11.3.x build

11.3.0 was failing

* try "11.3.1" since "11.3" didn't work

* added missing Python version

* Set build job name using matrix values

* tweak build job name

* ubuntu 11.3.1

* update cxx version to 20

* update cuda c++ standard to 20, cmake to 3.12

* moved cuda_standard 20 call, remove commented code

* removed cuda standard required for 20

* only c++17 is supported by nvcc

* add my old ci file, add entry to gitignore

* more complete build action from pr_build_action

* update required cmake version

* make c++ standard less restrictive, fix M_PI error

* update .gitignore

* Pass CUDA version directly to CMake

* Fix yml syntax error

* use toolset to specify cuda version

* switch to pass cuda path not version

* delete CMakeCace

* Fix configure cmake bash bug

* ignore error if CMakeCache does not exist

* run create package and upload result as artifact

* add script to make documentation for linux

* make DCE models default on

* attempt to cache CUDA install instead of build

* dumb fix

* add cache ID, "fix" cache path

* remove cache hit condition

somehow linux ran the windows install??????

* try fix cache condition

if cache hit, don't reinstall CUDA

* try fix cache path (real)

* add windows CUDA path to cache path

* try enabling separable compilation for windows

* the last commit sadly did not work

* add cached CUDA to PATH

* fix condition for last commit

(add CUDA path if cache hit)

* fix cache hit condition (real)

* try adding CUDA_PATH manually

* curse this esoteric syntax

* 🤡

* CUDA_PATH to env, bins to PATH

* specify path rather than use env var

env vars are not set until after the step that creates/updates them

* double \ for windows

pls work

* use bash for windows path set

* split base into major and minor, add 2 paths

paths are basically the same but may be needed for CMake detection

* fix dumb mistake

* try copy cuda extras into VS build customization

* try harder

* copy files, not directory

* fix syntax

* add -r to cp

* change cp dest to runner version of VS

* fix copy dest (real, hopefully)

* Remove Linux caching

Installing CUDA on Linux is fast enough without the cache.

* add MATLAB setup and Boost to CI

* split boost into OS specific steps, set BOOST_ROOT

may need to address ubuntu-22.04 problem -- can't seem to elegantly change version based on distribution

* forgot to add if os for boost

* add BOOST_ROOT env var to cmake config, add tests

* Disable performance tests + CUDA 10.1 job

idk what caused this dumb error outside of enabling our default models:  ptxas fatal   : Unresolved extern function '_Z3powfi'

* Delete c-cpp.yml

redundant with ci.yml

* make only examples & tests of included models

* revert to upload build dir

+++ examples/c++/Dixon_Fitting.cpp
@@ -123,7 +123,7 @@ void liver_fat_three()
 		for (int h =0; h < 7; h++)
 		{
 			// weight_list * e ^ (j * 2 * pi * ppm_list * TEn)
-			C_n += weight_list[h] * pow(expC, (j * 2.0f * piC * ppm_list[h] * TEnC[k]));
+			C_n += weight_list[h] * pow(expC, (j * (REAL)2.0 * piC * ppm_list[h] * TEnC[k]));
 		}
 
 		REAL x = user_info[k];
@@ -381,7 +381,7 @@ void liver_fat_four()
 		for (int h =0; h < 7; h++)
 		{
 			// weight_list * e ^ (j * 2 * pi * ppm_list * TEn)
-			C_n += weight_list[h] * pow(expC, (j * 2.0f * piC * ppm_list[h] * TEnC[k]));
+			C_n += weight_list[h] * pow(expC, (j * (REAL)2.0 * piC * ppm_list[h] * TEnC[k]));
 		}
 
 		REAL x = user_info[k];
